requirements.txt
numpy
opencv-python
PyQt5
scipy
ultralytics
Pillow
